Improve organization of test util packages
Some test utilities were previously in tests/integration/api, but they were also used in tests/e2e/operator. They should be in pkg/util/tests so that the e2e test package doesn't depend on the integration test package.

Additionally, test utilities that will ever only be used by e2e tests, should be in the e2e package.
