+++ pkg/test/util/ginkgo/ginkgo.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,18 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+package gutils
 
 import (
-	. "github.com/onsi/gomega"
-	gomegatypes "github.com/onsi/gomega/types"
+	g "github.com/onsi/ginkgo/v2"
+	"github.com/onsi/gomega"
+	"github.com/onsi/gomega/types"
 	"k8s.io/apimachinery/pkg/api/errors"
 )
 
-func ReturnNotFoundError() gomegatypes.GomegaMatcher {
-	return MatchError(errors.IsNotFound, "IsNotFound")
+// Success func will print the success message using the provided message
+// Arguments:
+// - string: message to be printed
+func Success(message string) {
+	g.GinkgoWriter.Println("* " + message)
 }
 
-func ReturnConflict() gomegatypes.GomegaMatcher {
-	return MatchError(errors.IsConflict, "IsConflict")
+func Step(text string, callback ...func()) {
+	g.GinkgoHelper()
+	g.By(text, callback...)
+}
+
+func ReturnNotFoundError() types.GomegaMatcher {
+	return gomega.MatchError(errors.IsNotFound, "IsNotFound")
 }
