--- pkg/util/tests/types/resourcecondition_types.go
@@ -1,29 +1,0 @@
-// Copyright Istio Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package resource
-
-type Condition struct {
-	Type   string `json:"type"`
-	Status string `json:"status"`
-}
-
-type Status struct {
-	Conditions []Condition `json:"conditions"`
-	Phase      string      `json:"phase"`
-}
-
-type Resource struct {
-	Status Status `json:"status"`
-}
