+++ tests/e2e/operator/env.go
