+++ tests/e2e/operator/util/helm/helm.go
@@ -18,8 +18,8 @@ import (
 	"fmt"
 	"strings"
 
-	g "github.com/istio-ecosystem/sail-operator/pkg/util/tests/ginkgo"
-	"github.com/istio-ecosystem/sail-operator/pkg/util/tests/shell"
+	g "github.com/istio-ecosystem/sail-operator/pkg/test/util/ginkgo"
+	"github.com/istio-ecosystem/sail-operator/tests/e2e/operator/util/shell"
 )
 
 // Install runs helm install in the given directory with params
