+++ tests/e2e/operator/util/kubectl/resourcelist_types.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,13 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package integration
+package kubectl
 
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
+type ResourceList struct {
+	APIVersion string        `json:"apiVersion"`
+	Items      []interface{} `json:"items"`
+	Kind       string        `json:"kind"`
+	Metadata   Metadata      `json:"metadata"`
+}
 
-func Step(text string, callback ...func()) {
-	GinkgoHelper()
-	By(text, callback...)
+type Metadata struct {
+	ResourceVersion string `json:"resourceVersion"`
 }
