+++ tests/e2e/operator/util/shell/shell.go
